+++ workers/ohsome_quality_analyst/api/api.py
@@ -27,7 +27,6 @@ from ohsome_quality_analyst.api.request_models import (
     IndicatorData,
     IndicatorDatabase,
     IndicatorEnum,
-    ProjectEnum,
     ReportBpolys,
     ReportDatabase,
     ReportEnum,
@@ -36,6 +35,7 @@ from ohsome_quality_analyst.api.request_models import (
 from ohsome_quality_analyst.api.response_models import (
     IndicatorMetadataResponse,
     MetadataResponse,
+    ProjectMetadataResponse,
     QualityDimensionMetadataResponse,
     ReportMetadataResponse,
     TopicMetadataResponse,
@@ -51,6 +51,11 @@ from ohsome_quality_analyst.definitions import (
     get_topic_definitions,
 )
 from ohsome_quality_analyst.geodatabase import client as db_client
+from ohsome_quality_analyst.projects.definitions import (
+    ProjectEnum,
+    get_project,
+    get_projects,
+)
 from ohsome_quality_analyst.quality_dimensions.definitions import (
     QualityDimensionEnum,
     get_quality_dimension,
@@ -148,10 +153,9 @@ class CustomJSONResponse(JSONResponse):
 
 
 @app.exception_handler(RequestValidationError)
-@app.exception_handler(ValidationError)
 async def validation_exception_handler(
-    request: Request,
-    exception: RequestValidationError | ValidationError,
+    _: Request,
+    exception: RequestValidationError,
 ):
     """Exception handler for validation exceptions."""
     return JSONResponse(
@@ -159,8 +163,8 @@ async def validation_exception_handler(
         content=jsonable_encoder(
             {
                 "apiVersion": __version__,
-                "detail": exception.errors(),
                 "type": "RequestValidationError",
+                "detail": exception.errors(),
             },
         ),
     )
@@ -172,24 +176,28 @@ async def validation_exception_handler(
 @app.exception_handler(RasterDatasetNotFoundError)
 @app.exception_handler(RasterDatasetUndefinedError)
 @app.exception_handler(SizeRestrictionError)
+@app.exception_handler(ValidationError)
 async def oqt_exception_handler(
-    request: Request,
-    exception: (
-        HexCellsNotFoundError
-        | TopicDataSchemaError
-        | OhsomeApiError
-        | RasterDatasetNotFoundError
-        | RasterDatasetUndefinedError
-        | SizeRestrictionError
-    ),
+    _: Request,
+    exception: HexCellsNotFoundError
+    | TopicDataSchemaError
+    | OhsomeApiError
+    | RasterDatasetNotFoundError
+    | RasterDatasetUndefinedError
+    | SizeRestrictionError
+    | ValidationError,
 ):
     """Exception handler for OQT exceptions."""
     return JSONResponse(
         status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
         content={
             "apiVersion": __version__,
-            "detail": exception.message,
             "type": exception.name,
+            "detail": [
+                {
+                    "msg": exception.message,
+                },
+            ],
         },
     )
 
@@ -334,6 +342,7 @@ async def metadata(project: ProjectEnum = DEFAULT_PROJECT) -> MetadataResponse:
     result = {
         "topics": get_topic_definitions(project=project.value),
         "quality-dimensions": get_quality_dimensions(),
+        "projects": get_projects(),
         "indicators": get_indicator_definitions(project=project.value),
         "reports": get_report_definitions(project=project.value),
     }
@@ -384,6 +393,26 @@ async def metadata_quality_dimension_by_key(
     return QualityDimensionMetadataResponse(
         result={key.value: get_quality_dimension(key.value)}
     )
+
+
+@app.get(
+    "/metadata/projects",
+    tags=["metadata"],
+)
+async def metadata_projects() -> ProjectMetadataResponse:
+    """Get projects."""
+    return ProjectMetadataResponse(result=get_projects())
+
+
+@app.get(
+    "/metadata/projects/{key}",
+    tags=["metadata"],
+)
+async def metadata_project_by_key(
+    key: ProjectEnum,
+) -> ProjectMetadataResponse:
+    """Get project by key."""
+    return ProjectMetadataResponse(result={key.value: get_project(key.value)})
 
 
 @app.get(